iOS: move runtime_attributes to ios/device
