+++ lib/calabash/ios.rb
@@ -14,7 +14,7 @@ module Calabash
       Calabash.send(:included, base)
     end
 
-    require 'calabash/ios/runtime_attributes'
+    require 'calabash/ios/device/runtime_attributes'
     require 'calabash/ios/routes/error'
     require 'calabash/ios/routes/route_mixin'
     require 'calabash/ios/routes/map_route'
