+++ lib/calabash/ios/device/runtime_attributes.rb
