+++ spec/lib/ios/device/runtime_attributes_spec.rb
